+++ demo-angular/references.d.ts
@@ -1,3 +1,3 @@
 /// <reference path="../src/node_modules/tns-platform-declarations/ios.d.ts" />
 /// <reference path="../src/node_modules/tns-platform-declarations/android.d.ts" />
-/// <reference path="../src/typings/java!uploadservice-3.4.2.ts" />
+/// <reference path="../src/typings/java!uploadservice-3.5.2.ts" />
