+++ src/platforms/android/include.gradle
@@ -4,5 +4,5 @@ android {
 
 //optional elements
 dependencies {
-    implementation 'net.gotev:uploadservice:3.4.2'
+    implementation 'net.gotev:uploadservice:3.5.2'
 }
